+++ src/Adapters/Laravel/Commands/TestCommand.php
@@ -4,15 +4,20 @@ declare(strict_types=1);
 
 namespace NunoMaduro\Collision\Adapters\Laravel\Commands;
 
-use Dotenv\Dotenv;
-use Dotenv\Repository\RepositoryBuilder;
+use Dotenv\Exception\InvalidPathException;
+use Dotenv\Parser\Parser;
+use Dotenv\Store\StoreBuilder;
 use Illuminate\Console\Command;
+use Illuminate\Support\Env;
 use Illuminate\Support\Str;
 use RuntimeException;
 use Symfony\Component\Process\Exception\ProcessSignaledException;
 use Symfony\Component\Process\Process;
 
-final class TestCommand extends Command
+/**
+ * @final
+ */
+class TestCommand extends Command
 {
     /**
      * The name and signature of the console command.
@@ -57,6 +62,15 @@ final class TestCommand extends Command
      */
     public function handle()
     {
+        if ((int) \PHPUnit\Runner\Version::id()[0] < 9) {
+            throw new RuntimeException('Running Collision ^5.0 artisan test command requires PHPUnit ^9.0.');
+        }
+
+        // @phpstan-ignore-next-line
+        if ((int) \Illuminate\Foundation\Application::VERSION[0] < 8) {
+            throw new RuntimeException('Running Collision ^5.0 artisan test command requires Laravel ^8.0.');
+        }
+
         $options = array_slice($_SERVER['argv'], $this->option('without-tty') ? 3 : 2);
 
         $this->clearEnv();
@@ -132,20 +146,45 @@ final class TestCommand extends Command
     protected function clearEnv()
     {
         if (!$this->option('env')) {
-            $repositories = RepositoryBuilder::create()
-                ->make();
-
-            $envs = Dotenv::create(
-                $repositories,
+            $vars = self::getEnvironmentVariables(
                 // @phpstan-ignore-next-line
                 $this->laravel->environmentPath(),
                 // @phpstan-ignore-next-line
                 $this->laravel->environmentFile()
-            )->safeLoad();
+            );
 
-            foreach (array_keys($envs) as $name) {
-                $repositories->clear($name);
+            $repository = Env::getRepository();
+
+            foreach ($vars as $name) {
+                $repository->clear($name);
             }
         }
+    }
+
+    /**
+     * @param string $path
+     * @param string $file
+     *
+     * @return array
+     */
+    protected static function getEnvironmentVariables($path, $file)
+    {
+        try {
+            $content = StoreBuilder::createWithNoNames()
+                ->addPath($path)
+                ->addName($file)
+                ->make()
+                ->read();
+        } catch (InvalidPathException $e) {
+            return [];
+        }
+
+        $vars = [];
+
+        foreach ((new Parser())->parse($content) as $entry) {
+            $vars[] = $entry->getName();
+        }
+
+        return $vars;
     }
 }